+++ src/IotHub/IotHub/Properties/Resources.Designer.cs
@@ -140,6 +140,15 @@ namespace Microsoft.Azure.Commands.Management.IotHub.Properties {
         }
         
         /// <summary>
+        ///   Looks up a localized string similar to ms-azure-iot-edge://.
+        /// </summary>
+        internal static string DEVICE_DEVICESCOPE_PREFIX {
+            get {
+                return ResourceManager.GetString("DEVICE_DEVICESCOPE_PREFIX", resourceCulture);
+            }
+        }
+        
+        /// <summary>
         ///   Looks up a localized string similar to Failover Iot Hub to the secondary location..
         /// </summary>
         internal static string ManualFailoverIotHub {