+++ src/IotHub/IotHub/ChangeLog.md
@@ -30,6 +30,9 @@
 	- Set-AzIotHubModule
 * Add cmdlet to get the connection string of a target IoT device in an Iot Hub.
 * Add cmdlet to get the connection string of a module on a target IoT device in an Iot Hub.
+* Add support to get/set parent device of an IoT device. New Cmdlets are:
+    - Get-AzIotHubDeviceParent
+    - Set-AzIotHubDeviceParent
 
 ## Version 2.1.0
 * Added support to manage devices in an Iot Hub. New Cmdlets are: