+++ src/IotHub/IotHub/ChangeLog.md
@@ -18,6 +18,25 @@
         - Additional information about change #1
 -->
 ## Upcoming Release
+* Add support to manage devices in an Iot Hub. New Cmdlets are:
+	- Add-AzIotHubDevice
+	- Get-AzIotHubDevice
+	- Remove-AzIotHubDevice
+	- Set-AzIotHubDevice
+* Add support to manage modules on a target Iot device in an Iot Hub. New Cmdlets are:
+	- Add-AzIotHubModule
+	- Get-AzIotHubModule
+	- Remove-AzIotHubModule
+	- Set-AzIotHubModule
+* Add cmdlet to get the connection string of a target IoT device in an Iot Hub.
+* Add cmdlet to get the connection string of a module on a target IoT device in an Iot Hub.
+
+## Version 2.1.0
+* Added support to manage devices in an Iot Hub. New Cmdlets are:
+	- `Add-AzIotHubDevice`
+	- `Get-AzIotHubDevice`
+	- `Remove-AzIotHubDevice`
+	- `Set-AzIotHubDevice`
 
 ## Version 2.0.1
 * Update references in .psd1 to use relative path